Fixes and general code clean-up
This PR targets the PHY classes timing issues and fixes an UL allocation problem that has been found during testing. On top of that, some dead code got removed and the documentation has been slightly improved.
In order to fix the synchronization issue, which originates from the fact that the OFDM symbols duration can not be represented as an integer in the Time's default unit of measure (nanoseconds), the various slots starting time is used as a sync anchor.
Specifically, each OFDM symbol lasts floor(SlotPeriod/SymbPerSlot) [ns] and at the end of the last symbol the remaining time to the actual next slot start is skipped by scheduling the StartSlot functions taking into account the exact slot period (therefore, recovering the ns that would be otherwise lost in the aforementioned approximation).
Regarding the UL allocation bug, I previously changed the timing of the refreshing of the slot allocation information. Turns out, such (apparently weird) timing did indeed make sense, as future slots might contain UL DCIs that are allocated in advance. Therefore, deleting them too much in advance causes problems: the allocation info needs to be re-initialized at specific instants, in order to preserve the UL allocations. The issue has been now fixed, and such timing sensitivity has been pointed out in the documentation.
Finally, I improved a bit the documentation and removed some small portions of code that I introduced but are not actually needed. I also added the # of retx to the traces (in order to visualize the retx traffic as different blocks in the GUI script)