Prevent triggering assert when using DFT-beamforming (#92)
Currently, the simulation crashes whenever DFT beamforming is used.
This PR fixes the above bug by reflecting the changes in the PhasedArrayModel APIs, which now requires beamforming vectors to be set before being retrieved. Previously, this was not happening when using DFT beamforming, as only the gNB antenna was configured.